+++ src/main/java/newsbot/ui/GeneralResponseButtons.java
@@ -1,6 +1,5 @@
-package newsbot.console;
+package newsbot.ui;
 import newsbot.news.NewsCategory;
-import newsbot.repository.UserProfileRepository;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardRow;
